Skybox added (29:26:33)
Fixed camera lock
Added cubemaps (for a skybox)
Added a set of cubes as a demo model to test engine

Importing models with transformations from their origins is currently having issues. Quick solve is to remove all transformations on origin and only transform the model within model space.
While I do believe that with a bit more time I could fix it, I've chosen to stop attempting to solve the problem and accept the quick fix. This choice was made after a couple hours of attempted debugging and I don't wish to contribute more time to the problem due to the target position being one of a gameplay engineer instead of a game engine programmer.

29 hours, 26 minutes and 33 seconds into the project